Remove opensearch feature flag
Removed `opensearch` feature.
Now both the code for Usearch and OpenSearch are compiled and available to use.

This change is required to build docker image able to use either provider.
One of the above is now chosen setting `SCYLLA_INDEXING_PROVIDER` environmental
variable to `opensearch` or `usearch` respectively.
By default the Usearch is being used.