validator/tests: fix ANN LIMIT tests
The LIMIT tests assumed that if there are enough data in the index
then the LIMIT will return all the data up to the limit.
According to the USearch documentation, that is not true and the LIMIT
is a parameter meaning that AT MOST the number of results will be returned.