Use cargo-machete for CI
The cargo-machete is a tool for checking if there are some unused dependencies
in crates. We shouldn't have unused dependencies in our source code. This
commit adds additional test in GitHub action to check this.

The commit adds also the `install-cargo-tools` script for installing in
development environment same tools used by the CI.