validator-tests: Enable tablets in tests
Switch from SimpleStrategy to NetworkTopologyStrategy to enable tablet
functionality, as tablets are not supported with SimpleStrategy.

Enable --rf-rack-valid-keyspaces to allow vector index creation on tables
with tablets. Without this flag, vector index creation fails with:
"Materialized views and secondary indexes are not supported on base tables
with tablets. To be able to use them, enable the configuration option
`rf_rack_valid_keyspaces`"